+++ packages/core/helpers/tests/testStandaloneBuild.ts
